Expresso and Unit tests

+++ app/src/main/java/com/example/testapp/data/StoryRepository.kt
@@ -1,8 +1,6 @@
 package com.example.testapp.data
 
-import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
-import androidx.lifecycle.Transformations
 import com.example.testapp.model.RoomResult
 import com.example.testapp.model.Stories
 import com.example.testapp.model.toRoomResult
@@ -11,6 +9,8 @@ import com.example.testapp.utils.API_KEY
 import io.reactivex.Observable
 import io.reactivex.android.schedulers.AndroidSchedulers
 import io.reactivex.schedulers.Schedulers
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.withContext
 import timber.log.Timber
 import javax.inject.Inject
 
@@ -30,15 +30,29 @@ class StoryRepository @Inject constructor(
                     .observeOn(AndroidSchedulers.mainThread())
                     .subscribe({ stories ->
                         val roomResults = stories.results.map { it.toRoomResult() }
-                        deleteAllStories()
-                        saveDataLocally(roomResults)
+                       // deleteAllStories()
+                       // saveDataLocally(roomResults)
                         storiesLiveData.value = roomResults
                     }, { error ->
                         Timber.e(error)
                     })
             )
         } else {
-            storiesLiveData.value = fetchDataFromRoom()
+
+           // storiesLiveData.value = fetchDataFromRoom()
+        }
+    }
+
+    suspend fun fetchDataCoroutine(isConnected: Boolean) {
+        withContext(Dispatchers.IO) {
+        if (isConnected) {
+                val stories = fetchDataFromNetworkCoroutine().results.map { it.toRoomResult() }
+                deleteAllStories()
+                saveDataLocally(stories)
+                storiesLiveData.postValue(stories)
+            }else{
+            storiesLiveData.postValue(fetchDataFromRoom())
+        }
         }
     }
 
@@ -46,15 +60,19 @@ class StoryRepository @Inject constructor(
         return apiService.getDataFromNetwork(API_KEY)
     }
 
-    private fun fetchDataFromRoom(): List<RoomResult> {
+    private suspend fun fetchDataFromNetworkCoroutine(): Stories {
+        return apiService.getFromNetworkCoroutine(API_KEY)
+    }
+
+    private suspend fun fetchDataFromRoom(): List<RoomResult> {
         return database.getStoryDao().getStories()
     }
 
-    private fun saveDataLocally(result: List<RoomResult>) {
+    private suspend fun saveDataLocally(result: List<RoomResult>) {
         database.getStoryDao().insertStory(result)
     }
 
-    private fun deleteAllStories() {
+    private suspend fun deleteAllStories() {
         database.getStoryDao()
             .deleteAllStories()
         Timber.d("Deleted")
